# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "MiraVelo Bike-Leasing API",
    "description" : "Customer-portal and back-office endpoints for the MiraVelo bike-leasing process. The engine-internal /engine-rest API is intentionally excluded.",
    "version" : "1.0"
  },
  "paths" : {
    "/api/bike-leasing" : {
      "get" : {
        "tags" : [ "list-leasing-applications-controller" ],
        "operationId" : "listLeasingApplications",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LeasingApplicationPageDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "submit-leasing-request-controller" ],
        "operationId" : "submit",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LeasingRequestInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LeasingApplicationCreatedDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "LeasingRequestInput" : {
        "type" : "object",
        "properties" : {
          "customerName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "age" : {
            "type" : "integer",
            "format" : "int32"
          },
          "monthlyNetIncome" : {
            "type" : "number",
            "format" : "double"
          },
          "bikeId" : {
            "type" : "string"
          },
          "bikeModel" : {
            "type" : "string"
          }
        },
        "required" : [ "age", "bikeId", "bikeModel", "customerName", "email", "monthlyNetIncome" ]
      },
      "LeasingApplicationCreatedDto" : {
        "type" : "object",
        "properties" : {
          "applicationId" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationId" ]
      },
...
      "LeasingApplicationPageDto" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LeasingApplicationSummaryDto"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "items", "page", "size", "totalElements", "totalPages" ]
      },
      "LeasingApplicationSummaryDto" : {
        "type" : "object",
        "properties" : {
          "applicationId" : {
            "type" : "string"
          },
          "customerName" : {
            "type" : "string"
          },
          "bikeId" : {
            "type" : "string"
          },
          "bikeModel" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "applicationId", "bikeId", "createdAt", "customerName", "status" ]
      },
...
    }
  }
}